pyproject.toml
[project]
name = "imei-check"
version = "0.1.0"
description = "test task 25.01.2025"
authors = [
    {name = "Dmitriy Boyarinov",email = "boyar.d@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pre-commit (>=4.1.0,<5.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "alembic (>=1.14.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
